Oi, so that would end up problematic after all.. \(facades always invisible and unable to interact properly...\)
